Move controls that have been converted to Blazor to their own namespace, for organization
